requirements.txt
# List your Python dependencies here
fastapi
uvicorn
diffusers
torch
transformers
Pillow
numpy
google-generativeai
huggingface_hub
tiktoken
accelerate
python-multipart
google-auth-oauthlib
